requirements.txt
matplotlib==3.7.1
matplotlib-inline==0.1.6
monotonic==1.5
numpy==1.24.3
open3d==0.17.0
opencv-contrib-python==3.4.8.29
opencv-python==4.5.1.48
openpyxl==3.0.3
pandas==1.1.2
Pillow==9.5.0
scikit-image==0.16.2
scikit-learn==1.2.2
scipy==1.10.1
tqdm==4.65.0

